Refactor: Extract ChatView Code to ConversationChatView
